+++ NEWS.md
@@ -6,6 +6,8 @@
 * `guess_repo()` now uses `gh::gh_tree_remote()` rather than gert - this eliminates the gert dependency. [#80]
 * `pb_release_delete()` introduced to delete existing releases. [#81]
 * `pb_new_release()` renamed to `pb_release_create()` to sync with the new delete function. 
+* Fix offer to create new release in `pb_upload()` - also switch to using `rlang::is_interactive()` to maybe one day test this.
+*
 
 # piggyback 0.1.3